requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0
django-admin-honeypot==1.1.0
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
openpyxl==3.0.9
psycopg2==2.9.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
